Bountiful Update
This update adds new items and equipment. Items now have three tiers, unlocking the higher tiers when you level up. Equipment now have two tiers, the high tier of each item will be unlocked as soon as you receive the first tier. Added the ability to confirm items and return to menu if you decide not to use. The item you first get will now be shower in the beginning of the story. Updated Ascii Art for Leviathan.